Extract Pagination logic from Grid into a new component (GridPaginator).
